requirements.txt
streamlit==1.37.0
pymupdf==1.24.9
llama-index-readers-file==0.1.32
llama-index-vector-stores-postgres==0.1.13
llama-index-embeddings-huggingface==0.2.2
llama-index-llms-llama-cpp==0.1.4
llama-cpp-python==0.2.85
psycopg2-binary==2.9.9
pgvector==0.2.5
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.31
greenlet==3.0.3